backend/requirements.txt
# Minimal requirements for demo
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
requests==2.31.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AI/NLP Libraries
transformers==4.36.2
torch==2.1.2
nltk==3.8.1
scikit-learn==1.3.2

# Data processing
pandas==2.1.4
numpy==1.24.3

# Email service
smtplib3==0.1.0
email-validator==2.1.0

# Kaggle API
kaggle==1.6.6

# Social media APIs
tweepy==4.14.0
linkedin-api==2.0.0

# Real-time updates
websockets==12.0
sse-starlette==1.8.2

# Task scheduling
apscheduler==3.10.4